updated: add base class NodeAdapter::Event
