+++ src/PlatformAdapter/Native/Sdk/Event/MouseEvent.hpp
@@ -9,13 +9,10 @@ namespace VGG
 namespace NodeAdapter
 {
 
-class MouseEvent : public BasicEvent<VGG::MouseEvent>
+class MouseEvent : public UIEvent<VGG::MouseEvent>
 {
 public:
   static void Init(napi_env env, napi_value exports);
-  static void Destructor(napi_env env, void* nativeObject, void* finalize_hint);
-
-  ~MouseEvent();
 
 private:
   static napi_value getModifierState(napi_env env, napi_callback_info info);
